+++ CHANGELOG.md
@@ -2,8 +2,22 @@
 
 ## Next
 
+## 1.8.0
+
+- Feature (`@grafana/faro-web-sdk`): track `web vitals` attribution (#595).
+- Feature (`@grafana/faro-web-sdk`): set span context for navigation events (#608).
+- Feature (`@grafana/faro-react`): add helper functions to initialize React Router integration (#622).
+
 - Enhancement (`@grafana/faro-web-sdk`): Auto extend a session if the Faro receiver indicates that a
   session is invalid (#591).
+- Enhancement (`@grafana/faro-web-tracing`): provide the `app.namespace` attribute in the app meta
+  which is attached as `service.namespace` to the resource attributes object (#627).
+
+- Dependencies (`@grafana/faro-web-tracing`): upgrade otel deps (#621).
+- Dependencies (`@grafana/faro-core`): upgrade otel deps (#621).
+- Dependencies (`@grafana/faro-transport-otlp-http [experimental]`): upgrade otel deps (#621).
+
+- Fix (`@grafana/faro-react`): Mark `react-router-dom` peer dependency as optional (#617).
 
 ## 1.7.3